simplify algopy benchmarks and add autograd benchmarks
